@@ -1,6 +1,7 @@
 {
   "name": "lwjgl.org",
-  "version": "1.1.0",
+  "type": "module",
+  "version": "1.2.0",
   "private": true,
   "license": "BSD-3-Clause",
   "description": "The website of LWJGL",
@@ -14,7 +15,7 @@
   },
   "homepage": "https://github.com/LWJGL/lwjgl3-www#readme",
   "dependencies": {
-    "aws-sdk": "2.577.0",
+    "aws-sdk": "2.578.0",
     "body-parser": "1.19.0",
     "chalk": "3.0.0",
     "compression": "1.7.4",
@@ -37,6 +38,7 @@
     "@babel/preset-env": "7.7.1",
     "@babel/preset-react": "7.7.0",
     "@babel/preset-typescript": "7.7.2",
+    "@fullhuman/postcss-purgecss": "1.3.0",
     "@prettier/plugin-pug": "1.0.10",
     "@reach/router": "1.3.0-beta.0",
     "@types/jszip": "3.1.6",
@@ -46,21 +48,20 @@
     "@types/scheduler": "0.16.1",
     "@typescript-eslint/eslint-plugin": "2.8.0",
     "@typescript-eslint/parser": "2.8.0",
-    "@webhotelier/webpack-fast-refresh": "1.0.9",
+    "@webhotelier/webpack-fast-refresh": "1.1.1",
     "autoprefixer": "9.7.2",
     "babel-loader": "8.0.6",
     "babel-plugin-emotion": "11.0.0-next.4",
     "babel-plugin-optimize-react": "0.0.4",
     "babel-plugin-transform-async-to-promises": "0.8.15",
     "babel-plugin-transform-react-remove-prop-types": "0.4.24",
-    "benchmark": "2.1.4",
     "bootstrap": "4.3.1",
     "cli-table": "0.3.1",
     "core-js": "3.4.2",
     "css-loader": "3.2.0",
     "cssnano": "4.1.10",
     "emotion": "11.0.0-next.0",
-    "eslint": "6.6.0",
+    "eslint": "6.7.0",
     "eslint-plugin-react-hooks": "2.3.0",
     "exports-loader": "0.7.0",
     "fbjs": "1.0.0",
