@@ -1,12 +1,15 @@
-'use strict';
-const path = require('path');
-const fs = require('fs').promises;
-const sass = require('node-sass');
-const postcss = require('postcss');
-const postcssConfig = require('../postcss.config.js');
-const { argv } = require('yargs');
+import path from 'path';
+import { promises as fs } from 'fs';
+import sass from 'node-sass';
+import postcss from 'postcss';
+import config from '../postcss.config.js';
+import yargs from 'yargs';
+import { promisify } from 'util';
 
-const renderSass = require('util').promisify(sass.render);
+import { fileURLToPath } from 'url';
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+const renderSass = promisify(sass.render);
 
 /*
   1. Render Sass to CSS
@@ -25,7 +28,7 @@ const process = async () => {
   const sourcePath = path.resolve(__dirname, '../client/styles/layout.scss');
   const targetPath = path.resolve(__dirname, '../public/css/core.css');
   const sourceMapPath = path.resolve(__dirname, '../public/css/core.css.map');
-  const SOURCEMAP = argv.sourcemap === true;
+  const SOURCEMAP = yargs.argv.sourcemap === true;
 
   console.log('Rendering Sass to CSS');
 
@@ -41,7 +44,7 @@ const process = async () => {
 
   console.log('Applying Autoprefixer & Minifying CSS');
 
-  const result = await postcss(postcssConfig.plugins).process(render.css, {
+  const result = await postcss(config.plugins).process(render.css, {
     from: sourcePath,
     to: targetPath,
     map: SOURCEMAP
