@@ -1,4 +1,1 @@
-'use strict';
-const prettyBytes = num => `${(num / 1000).toFixed(2)} kB`;
-
-module.exports = prettyBytes;
+export default num => `${(num / 1000).toFixed(2)} kB`;
